Move registries
