--- src/EntityRepository.js
@@ -1,1 +1,0 @@
-export default class EntityRepository {}
