+++ src/Registries/ComponentRegistry.js
@@ -1,4 +1,4 @@
-import Component from './Component';
+import Component from '../Component';
 
 export default class ComponentRegistry {
     #definitions = {};
